Refactor file interactions into a separate module for Obsidian plugin
- Extract file-related methods from chat_view.ts into a new interact_with_files.ts
- Create FileInteractions class to centralize file editing, parsing, and diff generation logic
- Improve modularity and separation of concerns in the Obsidian plugin
- Maintain existing functionality while making the code more organized and maintainable
